requirements.txt
requests
PyYAML
diskcache